backend/requirements.txt
fastapi
uvicorn
llama-index
llama-index-vector-stores-faiss
llama-index-embeddings-huggingface
transformers
torch
accelerate
faiss-cpu
pypdf
python-multipart
rank-bm25
networkx
spacy
python-dotenv